# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "App AI Gateway API",
    "version": "0.1.0",
    "description": "A multi-tenant, provider-native LLM proxy for mobile applications and trusted server backends."
  },
  "servers": [
    {
      "url": "https://gateway.example.com",
      "description": "Replace with your deployed gateway origin"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "ManagementBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "agw_mgmt_…",
        "description": "An organization-scoped management API key created by an owner or admin."
      },
      "OperatorSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "agw_operator_auth.session_token",
        "description": "The Better Auth operator session cookie. Admin requests also send x-console-request: 1."
      },
      "GatewayBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "A gateway access token, or an application API key for an issuer-less API-key app."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "data": {
                "type": "object",
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "number"
                    },
                    {
                      "type": "boolean"
                    }
                  ]
                },
                "description": "Machine-readable facts about this rejection, present only where the code alone is not actionable. A billing_request_quota_exceeded rejection carries limit, used, and resetAt; an app_rate_limited or app_budget_exhausted rejection carries scope, either user or app."
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "AppWrite": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9-]{0,62}$"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "config": {
            "$ref": "#/components/schemas/AppConfig"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "disabled"
            ]
          }
        },
        "required": [
          "name",
          "config"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/apps/{app}/auth/challenge": {
      "post": {
        "tags": [
          "Application authentication"
        ],
        "operationId": "createAppAttestChallenge",
        "summary": "Create an App Attest challenge",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "my-app"
            },
            "required": true,
            "name": "app",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "A five-minute, single-use challenge.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "challenge": {
                      "type": "string"
                    },
                    "expires_in": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "challenge",
                    "expires_in"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "No billing plan resolves for the organization.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated identity is not allowed to perform this operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/admin/apps": {
      "get": {
        "tags": [
          "Admin applications"
        ],
        "operationId": "listApps",
        "summary": "List applications",
        "security": [
          {
            "OperatorSession": []
          },
          {
            "ManagementBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Applications and current usage summaries.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "month": {
                      "type": "string"
                    },
                    "apps": {
                      "type": "array",
                      "items": {}
                    }
                  },
                  "required": [
                    "month",
                    "apps"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated identity is not allowed to perform this operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Admin applications"
        ],
        "operationId": "createApp",
        "summary": "Create an application",
        "description": "Omit `id` and the gateway assigns one from the name, suffixed to keep it unique across the deployment, and returns it as `app_id`. Send `id` and that exact id is used or the request is refused — a requested id is never silently replaced. API-key applications receive a one-time plaintext initial key in the response.",
        "security": [
          {
            "OperatorSession": []
          },
          {
            "ManagementBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppWrite"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Application created. `app_id` is the assigned id and the segment every gateway URL for this app uses.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "app_id": {
                      "type": "string"
                    },
                    "api_key": {}
                  },
                  "required": [
                    "app_id"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated identity is not allowed to perform this operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "`app_id_taken` — the requested `id` already belongs to an application. `invalid_request` — no unique generated id could be allocated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/admin/apps/{app}/validate": {
      "post": {
        "tags": [
          "Admin applications"
        ],
        "operationId": "validateApp",
        "summary": "Validate an application configuration without saving it",
        "security": [
          {
            "OperatorSession": []
          },
          {
            "ManagementBearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "my-app"
            },
            "required": true,
            "name": "app",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resolved valid configuration.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "The request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated identity is not allowed to perform this operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}